refactor: import 경로 수정 및 연락처 섹션 props 타입 동기화
