+++ apps/client/app/[exhibitionId]/artist/[artistId]/components/sections/BTSSection.tsx
@@ -1,6 +1,6 @@
 import { BTSCardGrid } from "@/components/common/Card/BTSCard/BTSCardGrid";
 import { Title } from "@/components/common/Title/Title";
-import type { ArtistDetail } from "../../api/artist-detail";
+import type { ArtistDetail } from "@/lib/api/artists/artist-detail.types";
 
 export function BTSSection({
 	exhibitionId,
